+++ common-core/src/main/java/com/goormgb/be/global/jwt/filter/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.goormgb.be.authguard.jwt.filter;
+package com.goormgb.be.global.jwt.filter;
 
 import java.io.IOException;
 import java.util.List;
@@ -10,9 +10,9 @@ import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
 
-import com.goormgb.be.authguard.jwt.enums.TokenType;
-import com.goormgb.be.authguard.jwt.provider.JwtTokenProvider;
-import com.goormgb.be.authguard.jwt.repository.AccessTokenBlacklistRepository;
+import com.goormgb.be.global.jwt.enums.TokenType;
+import com.goormgb.be.global.jwt.provider.JwtTokenProvider;
+import com.goormgb.be.global.jwt.repository.AccessTokenBlacklistRepository;
 import com.goormgb.be.global.exception.CustomException;
 import com.goormgb.be.global.exception.ErrorCode;
 import com.goormgb.be.global.support.Preconditions;
